--- a/src/Nethermind.Int256.Benchmark/Benchmarks.cs
+++ b/src/Nethermind.Int256.Benchmark/Benchmarks.cs
@@ -89,7 +89,7 @@ namespace Nethermind.Int256.Benchmark
         public (int, Int256) D;
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -107,7 +107,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddSigned : SignedTwoParamBenchmarkBase
     {
@@ -125,7 +125,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -143,7 +143,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractSigned : SignedTwoParamBenchmarkBase
     {
@@ -161,7 +161,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddModUnsinged : UnsignedThreeParamBenchmarkBase
     {
@@ -179,7 +179,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddModSinged : SignedThreeParamBenchmarkBase
     {
@@ -197,7 +197,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractModUnsinged : UnsignedThreeParamBenchmarkBase
     {
@@ -215,7 +215,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractModSigned : SignedThreeParamBenchmarkBase
     {
@@ -233,7 +233,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplyUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -251,7 +251,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplySigned : SignedTwoParamBenchmarkBase
     {
@@ -269,7 +269,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplyModUnsigned : UnsignedThreeParamBenchmarkBase
     {
@@ -287,7 +287,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplyModSigned : SignedThreeParamBenchmarkBase
     {
@@ -305,7 +305,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class DivideUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -323,7 +323,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class DivideSigned : SignedTwoParamBenchmarkBase
     {
@@ -341,7 +341,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpUnsigned : UnsignedIntTwoParamBenchmarkBase
     {
@@ -359,7 +359,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpSigned : SignedIntTwoParamBenchmarkBase
     {
@@ -377,7 +377,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpModUnsigned : UnsignedThreeParamBenchmarkBase
     {
@@ -395,7 +395,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpModSigned : SignedBenchmarkBase
     {
@@ -422,7 +422,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class LeftShiftUnsigned : UnsignedIntTwoParamBenchmarkBase
     {
@@ -440,7 +440,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class LeftShiftSigned : SignedIntTwoParamBenchmarkBase
     {
@@ -458,7 +458,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class RightShiftUnsigned : UnsignedIntTwoParamBenchmarkBase
     {
@@ -476,7 +476,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class RightShiftSigned : SignedIntTwoParamBenchmarkBase
     {
@@ -494,7 +494,7 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net60)]
+    [SimpleJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class IsZeroOne
     {