--- a/src/Nethermind.Int256.Benchmark/Benchmarks.cs
+++ b/src/Nethermind.Int256.Benchmark/Benchmarks.cs
@@ -89,7 +89,8 @@ namespace Nethermind.Int256.Benchmark
         public (int, Int256) D;
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -107,7 +108,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddSigned : SignedTwoParamBenchmarkBase
     {
@@ -125,7 +127,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -143,7 +146,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractSigned : SignedTwoParamBenchmarkBase
     {
@@ -161,7 +165,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddModUnsinged : UnsignedThreeParamBenchmarkBase
     {
@@ -179,7 +184,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class AddModSinged : SignedThreeParamBenchmarkBase
     {
@@ -197,7 +203,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractModUnsinged : UnsignedThreeParamBenchmarkBase
     {
@@ -215,7 +222,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class SubtractModSigned : SignedThreeParamBenchmarkBase
     {
@@ -233,7 +241,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplyUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -251,7 +260,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplySigned : SignedTwoParamBenchmarkBase
     {
@@ -269,7 +279,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplyModUnsigned : UnsignedThreeParamBenchmarkBase
     {
@@ -287,7 +298,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class MultiplyModSigned : SignedThreeParamBenchmarkBase
     {
@@ -305,7 +317,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class DivideUnsigned : UnsignedTwoParamBenchmarkBase
     {
@@ -323,7 +336,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class DivideSigned : SignedTwoParamBenchmarkBase
     {
@@ -341,7 +355,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpUnsigned : UnsignedIntTwoParamBenchmarkBase
     {
@@ -359,7 +374,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpSigned : SignedIntTwoParamBenchmarkBase
     {
@@ -377,7 +393,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpModUnsigned : UnsignedThreeParamBenchmarkBase
     {
@@ -395,7 +412,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class ExpModSigned : SignedBenchmarkBase
     {
@@ -422,7 +440,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class LeftShiftUnsigned : UnsignedIntTwoParamBenchmarkBase
     {
@@ -440,7 +459,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class LeftShiftSigned : SignedIntTwoParamBenchmarkBase
     {
@@ -458,7 +478,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class RightShiftUnsigned : UnsignedIntTwoParamBenchmarkBase
     {
@@ -476,7 +497,8 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class RightShiftSigned : SignedIntTwoParamBenchmarkBase
     {
@@ -494,11 +516,12 @@ namespace Nethermind.Int256.Benchmark
         }
     }
 
-    [SimpleJob(RuntimeMoniker.Net70)]
+    [SimpleJob(RuntimeMoniker.Net70, baseline: true)]
+    [NoIntrinsicsJob(RuntimeMoniker.Net70)]
     [MemoryDiagnoser]
     public class IsZeroOne
     {
-        public UInt256[] Values { get; } = {UInt256.Zero, UInt256.One, UInt256.MaxValue};
+        public UInt256[] Values { get; } = { UInt256.Zero, UInt256.One, UInt256.MaxValue };
 
         [ParamsSource(nameof(Values))]
         public UInt256 A;
@@ -508,7 +531,7 @@ namespace Nethermind.Int256.Benchmark
         {
             return A.IsZero;
         }
-        
+
         [Benchmark]
         public bool IsOne()
         {