--- a/src/Nethermind.Int256.Benchmark/Nethermind.Int256.Benchmark.csproj
+++ b/src/Nethermind.Int256.Benchmark/Nethermind.Int256.Benchmark.csproj
@@ -2,11 +2,11 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+      <PackageReference Include="BenchmarkDotNet" Version="0.13.3" />
     </ItemGroup>
 
     <ItemGroup>