--- a/src/Nethermind.Int256/Nethermind.Int256.csproj
+++ b/src/Nethermind.Int256/Nethermind.Int256.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <Nullable>enable</Nullable>
         <LangVersion>latest</LangVersion>
         <Deterministic>true</Deterministic>