--- a/src/Nethermind.Int256/UInt256.cs
+++ b/src/Nethermind.Int256/UInt256.cs
@@ -823,19 +823,42 @@ namespace Nethermind.Int256
         // Subtract sets res to the difference a-b
         public static void Subtract(in UInt256 a, in UInt256 b, out UInt256 res)
         {
-            ulong carry = 0ul;
-            SubtractWithBorrow(a.u0, b.u0, ref carry, out ulong res0);
-            SubtractWithBorrow(a.u1, b.u1, ref carry, out ulong res1);
-            SubtractWithBorrow(a.u2, b.u2, ref carry, out ulong res2);
-            SubtractWithBorrow(a.u3, b.u3, ref carry, out ulong res3);
-            res = new UInt256(res0, res1, res2, res3);
+            if (Avx2.IsSupported)
+            {
+                var av = Unsafe.As<UInt256,Vector256<ulong>>(ref Unsafe.AsRef(in a));
+                var bv = Unsafe.As<UInt256,Vector256<ulong>>(ref Unsafe.AsRef(in b));
 
-                // Lookup the carries to broadcast to the Vectors
+                var result = Avx2.Subtract(av, bv);
+                // Invert top bits as Avx2.CompareGreaterThan is only available for longs, not unsigned
+                var resultSigned = Avx2.Xor(result, Vector256.Create<ulong>(0x8000_0000_0000_0000));
+                var avSigned = Avx2.Xor(av, Vector256.Create<ulong>(0x8000_0000_0000_0000));
+
+                // Which vectors need to borrow from the next
+                var vBorrow = Avx2.CompareGreaterThan(Unsafe.As<Vector256<ulong>, Vector256<long>>(ref resultSigned),
+                                                      Unsafe.As<Vector256<ulong>, Vector256<long>>(ref avSigned));
+
+                // Move borrow from Vector space to int
+                var borrow = Avx.MoveMask(Unsafe.As<Vector256<long>, Vector256<double>>(ref vBorrow));
+
+                // All zeros will cascade another borrow when borrow is subtracted from it
+                var vCascade = Avx2.CompareEqual(result, Vector256<ulong>.Zero);
+                // Move cascade from Vector space to int
+                var cascade = Avx.MoveMask(Unsafe.As<Vector256<ulong>, Vector256<double>>(ref vCascade));
+
+                // Use ints to work out the Vector cross lane cascades
+                // Move borrow to next bit and add cascade
+                borrow = cascade + 2 * borrow; // lea
+                // Remove cascades not effected by borrow
+                cascade ^= borrow;
+                // Choice of 16 vectors
+                cascade &= 0x0f;
+
+                // Lookup the borrows to broadcast to the Vectors
                 var cascadedBorrows = Unsafe.Add(ref Unsafe.As<byte, Vector256<ulong>>(ref MemoryMarshal.GetReference(s_broadcastLookup)), cascade);
 
                 // Mark res as initalized so we can use it as left said of ref assignment
                 Unsafe.SkipInit(out res);
-                // Add the cascadedCarries to the result
+                // Subtract the cascadedBorrows from the result
                 Unsafe.As<UInt256,Vector256<ulong>>(ref res) = Avx2.Subtract(result, cascadedBorrows);
             }
             else