… .github/workflows/build-solution.yml
… .github/workflows/run-tests.yml
… src/Nethermind.Int256.Test/Nethermind.Int256.Test.csproj