|              Method | EnvironmentVariables |     A |     B |      Mean | Ratio |
|-------------------- |--------------------- |------ |------ |----------:|------:|
| Subtract_BigInteger |                Empty | [160] | [160] | 43.982 ns |  1.00 |
|    Subtract_UInt256 |  EnableHWIntrinsic=0 | [160] | [160] |  8.675 ns |  0.20 |
|    Subtract_UInt256 |                 AVX2 | [160] | [160] |  3.230 ns |  0.07 |
|                     |                      |       |       |           |       |
| Subtract_BigInteger |                Empty | [160] | [156] | 41.430 ns |  1.00 |
|    Subtract_UInt256 |  EnableHWIntrinsic=0 | [160] | [156] |  8.044 ns |  0.20 |
|    Subtract_UInt256 |                 AVX2 | [160] | [156] |  3.235 ns |  0.08 |
|                     |                      |       |       |           |       |
| Subtract_BigInteger |                Empty | [156] | [160] | 40.099 ns |  1.00 |
|    Subtract_UInt256 |  EnableHWIntrinsic=0 | [156] | [160] |  8.118 ns |  0.20 |
|    Subtract_UInt256 |                 AVX2 | [156] | [160] |  3.152 ns |  0.08 |
|                     |                      |       |       |           |       |
| Subtract_BigInteger |                Empty | [156] | [156] | 43.780 ns |  1.00 |
|    Subtract_UInt256 |  EnableHWIntrinsic=0 | [156] | [156] |  8.027 ns |  0.18 |
|    Subtract_UInt256 |                 AVX2 | [156] | [156] |  3.328 ns |  0.08 |

|         Method | EnvironmentVariables |     A |     B |      Mean | Ratio |
|--------------- |--------------------- |------ |------ |----------:|------:|
| Add_BigInteger |                Empty | [160] | [160] | 83.919 ns |  1.00 |
|    Add_UInt256 |  EnableHWIntrinsic=0 | [160] | [160] |  8.040 ns |  0.10 |
|    Add_UInt256 |                 AVX2 | [160] | [160] |  3.493 ns |  0.04 |
|                |                      |       |       |           |       |
| Add_BigInteger |                Empty | [160] | [156] | 82.518 ns |  1.00 |
|    Add_UInt256 |  EnableHWIntrinsic=0 | [160] | [156] |  8.154 ns |  0.10 |
|    Add_UInt256 |                 AVX2 | [160] | [156] |  5.307 ns |  0.06 |
|                |                      |       |       |           |       |
| Add_BigInteger |                Empty | [156] | [160] | 85.307 ns |  1.00 |
|    Add_UInt256 |  EnableHWIntrinsic=0 | [156] | [160] |  8.086 ns |  0.09 |
|    Add_UInt256 |                 AVX2 | [156] | [160] |  5.373 ns |  0.06 |
|                |                      |       |       |           |       |
| Add_BigInteger |                Empty | [156] | [156] | 39.583 ns |  1.00 |
|    Add_UInt256 |  EnableHWIntrinsic=0 | [156] | [156] |  8.193 ns |  0.21 |
|    Add_UInt256 |                 AVX2 | [156] | [156] |  3.155 ns |  0.08 |

/cc @LukaszRozmej