@LukaszRozmej 2023-01-14
Overall great. But the question that comes to mind is this the optimal way, or better to go with Int128/UInt128?