@benaadams 2023-01-14
In .NET 7 UInt128 doesn't use intrinsics yet and and is only a general software implementation. It would also require twice as many to do 256 bit (x 2 128). Not sure 128bit intrinsics would break even vs non intrinsics as Avx2 is only twice as fast not four times.


Will add proposal for UInt256 :)