@LukaszRozmej 2023-01-14
Yeah made a quick check. For now it looks like `Uint128` it is slower than `ulong` version