@benaadams 2023-01-15
> Will add proposal for UInt256 :)

https://github.com/dotnet/runtime/issues/80663