@LukaszRozmej 2023-01-16
I doubt it will go through. We are happy for this repo to become THE Int256 library for .Net.